pyproject.toml
[tool.poetry]
name = "block-streamer"
version = "0.1.0"
description = "Naive Ethereum block streamer using providers"
authors = ["Moshe sar idan <msi@dev.com>"]
readme = "README.md"
packages = [{include = "block_streamer"}]

[tool.poetry.dependencies]
python = ">=3.10,<4"
web3 = ">=6.15.1"
pyyaml = ">=6.0"
pydantic = ">=2.0.0"
python-dotenv = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
black = ">=24.2.0"
ruff = ">=0.3.0"
mypy = ">=1.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.isort]
known-first-party = ["block_streamer"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
